Milestone point for Simulator workloads.  The 'WithLobbyReturn' bundle encpsulates a FInite State Machine that defines a pattern of user behavior.  It is a building block that is attached to a simulation with parmeters bound that define things like the floors its Travellers target, and the arrival rate of their entry into the simulation (a Poisson Process).  This is the first of a pattern, which will see additional subtypes to build out a repetoire of simulated passenger behaviors over time.
There is currently a bug in the implementation related to how its Finite State Machine is constructed.  Travellers from different Populations (different configuration values for behavior parmeters) need to have diffeent FInite State Machines, although each machine is constructed by a common recipe.  The WithLobbyReturn behavior requires on additional State not present in the core abstract traveller concept.  Unfortunately, this state is only getting created one time and is incorrectly shared by multiple Population's FSMs.  The fix is not that convoluted, but I want to checkpoint this as it is for safety's sake just in case.